refactor(chatgpt): split image transformation modules
