+++ litellm/__init__.py
@@ -1871,8 +1871,10 @@ if TYPE_CHECKING:
     from .llms.chatgpt.responses.transformation import (
         ChatGPTResponsesAPIConfig as ChatGPTResponsesAPIConfig,
     )
-    from .llms.chatgpt.image_generation.transformation import (
+    from .llms.chatgpt.image_edit import (
         ChatGPTImageEditConfig as ChatGPTImageEditConfig,
+    )
+    from .llms.chatgpt.image_generation import (
         ChatGPTImageGenerationConfig as ChatGPTImageGenerationConfig,
     )
     from .llms.gigachat.chat.transformation import GigaChatConfig as GigaChatConfig
