+++ tests/test_litellm/llms/chatgpt/chatgpt_image_test_utils.py
@@ -1,0 +1,10 @@
+from typing import Any
+
+
+class MockLogging:
+    def post_call(self, *args, **kwargs):
+        pass
+
+
+def mock_logging() -> Any:
+    return MockLogging()
